# public/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"UserDTO": {
				"properties": {
					"email": {
						"type": "string"
					},
					"firstName": {
						"type": "string"
					},
					"lastName": {
						"type": "string"
					},
					"age": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"email",
					"firstName",
					"lastName",
					"age"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IErrorHandler": {
				"properties": {
					"statusCode": {
						"type": "number",
						"format": "double"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"statusCode",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserCreateDTO": {
				"properties": {
					"email": {
						"type": "string"
					},
					"firstName": {
						"type": "string"
					},
					"lastName": {
						"type": "string"
					},
					"age": {
						"type": "number",
						"format": "double"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"email",
					"firstName",
					"lastName",
					"age",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AuthDTO": {
				"properties": {
					"token": {
						"type": "string"
					}
				},
				"required": [
					"token"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AuthRequestDTO": {
				"properties": {
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"email",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"MovieDetailDTO": {
				"properties": {
					"title": {
						"type": "string"
					},
					"synopsis": {
						"type": "string"
					},
					"director": {
						"type": "string"
					},
					"genre": {
						"type": "string"
					},
					"actors": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"rating": {
						"type": "string"
					}
				},
				"required": [
					"title",
					"director",
					"genre",
					"actors",
					"rating"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ActorDTO": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					}
				},
				"required": [
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"MovieDTO": {
				"properties": {
					"title": {
						"type": "string"
					},
					"synopsis": {
						"type": "string"
					},
					"director": {
						"type": "string"
					},
					"genre": {
						"type": "string"
					},
					"actors": {
						"items": {
							"$ref": "#/components/schemas/ActorDTO"
						},
						"type": "array"
					}
				},
				"required": [
					"title",
					"director",
					"genre",
					"actors"
				],
				"type": "object",
				"additionalProperties": false
			},
			"MovieCreateDTO": {
				"description": "DTO used for creation of movies.\nActors will be translated from String array to an array of objects containing the actor name",
				"properties": {
					"title": {
						"type": "string"
					},
					"synopsis": {
						"type": "string"
					},
					"director": {
						"type": "string"
					},
					"genre": {
						"type": "string"
					},
					"actors": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"title",
					"director",
					"genre",
					"actors"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RateRequestDTO": {
				"properties": {
					"rating": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"rating"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"jwt": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT",
				"scopes": [
					"admin",
					"user"
				]
			}
		}
	},
	"info": {
		"title": "movie_rest_api",
		"version": "0.0.1",
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/user/{id}": {
			"get": {
				"operationId": "GetUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserDTO"
								}
							}
						}
					},
					"400": {
						"description": "Bad id",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IErrorHandler"
								}
							}
						}
					},
					"404": {
						"description": "Not Found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IErrorHandler"
								}
							}
						}
					}
				},
				"description": "Retrieves user by id",
				"tags": [
					"User"
				],
				"security": [
					{
						"jwt": [
							"User"
						]
					}
				],
				"parameters": [
					{
						"description": "User id",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"delete": {
				"operationId": "DeleteUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "boolean"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized operation",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IErrorHandler"
								}
							}
						}
					},
					"404": {
						"description": "Not Found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IErrorHandler"
								}
							}
						}
					}
				},
				"description": "Excludes user from system",
				"tags": [
					"User"
				],
				"security": [
					{
						"jwt": [
							"User",
							"Admin"
						]
					}
				],
				"parameters": [
					{
						"description": "user id to exclude",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserDTO"
								}
							}
						}
					},
					"400": {
						"description": "Invalid Data",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IErrorHandler"
								}
							}
						}
					},
					"404": {
						"description": "Not Found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IErrorHandler"
								}
							}
						}
					}
				},
				"description": "Updates information for an user",
				"tags": [
					"User"
				],
				"security": [
					{
						"jwt": [
							"User"
						]
					}
				],
				"parameters": [
					{
						"description": "user id to update",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"description": "user data to be updated",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UserDTO"
							}
						}
					}
				}
			}
		},
		"/user/register": {
			"post": {
				"operationId": "RegisterUser",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserDTO"
								}
							}
						}
					},
					"400": {
						"description": "Invalid data",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IErrorHandler"
								}
							}
						}
					}
				},
				"description": "Creates user",
				"tags": [
					"User"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"description": "Formatted user data to be created",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UserCreateDTO"
							}
						}
					}
				}
			}
		},
		"/admin/{id}": {
			"get": {
				"operationId": "GetAdmin",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserDTO"
								}
							}
						}
					},
					"400": {
						"description": "Bad id",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IErrorHandler"
								}
							}
						}
					},
					"404": {
						"description": "Not Found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IErrorHandler"
								}
							}
						}
					}
				},
				"description": "Retrieves an admin by id",
				"tags": [
					"Admin"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"delete": {
				"operationId": "DeleteAdmin",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "boolean"
								}
							}
						}
					},
					"404": {
						"description": "Not Found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IErrorHandler"
								}
							}
						}
					}
				},
				"description": "Excludes admin from system\nAuthenticated Endpoint: Admins",
				"tags": [
					"Admin"
				],
				"security": [
					{
						"jwt": [
							"Admin"
						]
					}
				],
				"parameters": [
					{
						"description": "user id to exclude",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateAdmin",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserDTO"
								}
							}
						}
					},
					"400": {
						"description": "Invalid data",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IErrorHandler"
								}
							}
						}
					},
					"404": {
						"description": "Not Found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IErrorHandler"
								}
							}
						}
					}
				},
				"description": "Updates information for an admin\nAuthenticated Endpoint: Admins",
				"tags": [
					"Admin"
				],
				"security": [
					{
						"jwt": [
							"Admin"
						]
					}
				],
				"parameters": [
					{
						"description": "user id to update",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"description": "user data to be updated",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UserDTO"
							}
						}
					}
				}
			}
		},
		"/admin/register": {
			"post": {
				"operationId": "RegisterAdmin",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserDTO"
								}
							}
						}
					},
					"400": {
						"description": "Invalid data",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IErrorHandler"
								}
							}
						}
					}
				},
				"description": "Register a new administrator",
				"tags": [
					"Admin"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"description": "Body of the user to be created as admin",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UserCreateDTO"
							}
						}
					}
				}
			}
		},
		"/auth": {
			"post": {
				"operationId": "Authenticate",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AuthDTO"
								}
							}
						}
					},
					"400": {
						"description": "Invalid data",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IErrorHandler"
								}
							}
						}
					}
				},
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AuthRequestDTO"
							}
						}
					}
				}
			}
		},
		"/movie/{id}": {
			"get": {
				"operationId": "GetMovieDetails",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MovieDetailDTO"
								}
							}
						}
					},
					"400": {
						"description": "Bad id",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IErrorHandler"
								}
							}
						}
					},
					"404": {
						"description": "Not Found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IErrorHandler"
								}
							}
						}
					}
				},
				"description": "Retrieves movie detail by id",
				"tags": [
					"Movie"
				],
				"security": [],
				"parameters": [
					{
						"description": "User id",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/movie": {
			"post": {
				"operationId": "RegisterMovie",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MovieDTO"
								}
							}
						}
					},
					"400": {
						"description": "Invalid Data",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IErrorHandler"
								}
							}
						}
					}
				},
				"description": "Register a movie in the system.\nAuthenticated Endpoint: Admins",
				"tags": [
					"Movie"
				],
				"security": [
					{
						"jwt": [
							"Admin"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"description": "- Formatted user data to be created",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/MovieCreateDTO"
							}
						}
					}
				}
			}
		},
		"/movie/{id}/rate": {
			"post": {
				"operationId": "RegisterRate",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RateRequestDTO"
								}
							}
						}
					},
					"400": {
						"description": "Invalid Data",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IErrorHandler"
								}
							}
						}
					},
					"404": {
						"description": "Not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IErrorHandler"
								}
							}
						}
					}
				},
				"description": "Rate a movie.\nAuthenticated method: Users",
				"tags": [
					"Movie"
				],
				"security": [
					{
						"jwt": [
							"User"
						]
					}
				],
				"parameters": [
					{
						"description": "- Movie id being rated",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"description": "- Payload with the value of the rating",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RateRequestDTO"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}